Backend/requirements.txt
# ===============================
# Core Django & REST Framework
# ===============================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
asgiref
sqlparse
whitenoise

# ===============================
# API Documentation & Environment
# ===============================
drf-yasg==1.21.7
python-decouple==3.8
python-dotenv

# ===============================
# Task Queue & Caching
# ===============================
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# ===============================
# File & Data Handling
# ===============================
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# ===============================
# PDF & Report Generation
# ===============================
reportlab==4.0.7
# weasyprint==60.1  # Optional for advanced PDF generation

# ===============================
# Data Analysis & Visualization
# ===============================
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# ===============================
# AI / ML & Language Processing
# ===============================
scikit-learn
scipy
torch
transformers
nltk
joblib
regex
tqdm
faiss-cpu
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith
google-generativeai
google-ai-generativelanguage
huggingface-hub
safetensors

# ===============================
# Google & Cloud Dependencies
# ===============================
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
grpcio
grpcio-status
proto-plus
protobuf
httplib2
cachetools
rsa

# ===============================
# Utilities & JSON Handling
# ===============================
attrs
dataclasses-json
marshmallow
typing-inspect
typing_extensions
pydantic
pydantic_core
annotated-types
mypy-extensions
orjson
jsonpatch
jsonpointer
pyparsing
packaging
MarkupSafe
Jinja2
PyYAML
colorama
click
filelock

# ===============================
# Database & ORM
# ===============================
SQLAlchemy
greenlet
psycopg

# ===============================
# Audio, Speech, & Text Processing
# ===============================
SpeechRecognition
pyttsx3
PyPDF2
comtypes

# ===============================
# Network & HTTP Libraries
# ===============================
aiohttp
aiosignal
frozenlist
multidict
charset-normalizer
urllib3
idna
uritemplate
yarl

# ===============================
# Math, Scientific, Misc
# ===============================
sympy
mpmath
networkx
mkl
tenacity
threadpoolctl
